[FE] user-context のリファクタ
