+++ frontend/app/routes/client._index/driving-state/pickup.tsx
@@ -6,11 +6,10 @@ import { LocationButton } from "~/components/modules/location-button/location-bu
 import { ModalHeader } from "~/components/modules/modal-header/modal-header";
 import { Price } from "~/components/modules/price/price";
 import { Text } from "~/components/primitives/text/text";
-import { useUserContext } from "~/contexts/user-context";
+import { useClientContext } from "~/contexts/client-context";
 
 export const Pickup: FC = () => {
-  const { payload = {} } = useUserContext();
-  const { chair, fare, coordinate } = payload;
+  const { data } = useClientContext();
   return (
     <div className="w-full h-full flex flex-col items-center justify-center max-w-md mx-auto">
       <ModalHeader title="目的地に向けて出発します" subTitle="お座りください">
@@ -21,28 +20,30 @@ export const Pickup: FC = () => {
             className="absolute left-0 bottom-[-5px]"
           />
           <ChairIcon
-            model={chair?.model ?? ""}
+            model={data?.chair?.model ?? ""}
             width={60}
             height={60}
             className="absolute bottom-0 right-0 animate-shake"
           />
         </div>
       </ModalHeader>
-      {chair && <ChairInformation className="mb-8" chair={chair} />}
+      {data?.chair && <ChairInformation className="mb-8" chair={data.chair} />}
       <LocationButton
         label="現在地"
-        location={coordinate?.pickup}
+        location={data?.pickup_coordinate}
         className="w-full"
         disabled
       />
       <Text size="xl">↓</Text>
       <LocationButton
         label="目的地"
-        location={coordinate?.destination}
+        location={data?.destination_coordinate}
         className="w-full"
         disabled
       />
-      {fare && <Price pre="運賃" value={fare} className="mt-8"></Price>}
+      {data?.fare && (
+        <Price pre="運賃" value={data?.fare} className="mt-8"></Price>
+      )}
     </div>
   );
 };
